Markdownの変換は、ReadingClubのViewで使うものなのでHelper書く場所も変更した
